% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeGRM_man.R
\name{$computeGRM}
\alias{$computeGRM}
\title{GRM method: Construct a genomic relationship matrix (GRM)}
\usage{
GRMobj$computeGRM(name, method="VanRaden", ep=0, snpsubset=NULL,
                  filter=list(MAF=NULL, MISS=NULL, PVALUE=NULL), ...)
}
\arguments{
\item{name}{A character string giving the name of the GRM analysis.}

\item{method}{A character string specifying whether the VanRaden (\code{'VanRaden'}) based estimator or
the Weir-Goudet (\code{'WG'}) estimator is used to construct the GRM.}

\item{ep}{Sequencing error value. Can be a single number (sequencing error rate same for all genotypes),
a vector equal to the number of SNPs (SNP specific sequencing error rate) or a matrix the same diminsion as the data (SNP and
individual specific sequencing error rate).}

\item{snpsubset}{Vector of indices of SNPs that should be retained in the analysis. Useful for subsetting the SNPs before
any filtering is applied.}

\item{filter}{Named list of thresholds for various filtering criteria.
See below for details.}

\item{...}{Not used currently.}
}
\description{
Method for constructing a genomic relationship matrix (GRM) for a diploid or autopolyploid population.
}
\details{
The filtering criteria currently implemented are:
\itemize{
\item{Minor allele frequency (\code{MAF}): }{SNPs are discarded if their MAF is less than the threshold (default is \code{NULL})}
\item{Proportion of missing data (\code{MISS}): }{SNPs are discarded if the proportion of individuals with no reads (e.g. missing genotype)
 is greater than the threshold value (default is \code{NULL}).}
\item{h (\code{PVALUE}): }{}
}
If a filtering criteria is set to \code{NULL}, then no filtering in regard to
that threshold is applied.
}
\seealso{
\code{\link{GRM}}
}
\author{
Timothy P. Bilton
}
