% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeGRM.R
\name{computeGRM}
\alias{computeGRM}
\title{GRM method: Construct a genomic relationship matrix (GRM)}
\usage{
computeGRM(
  ref,
  alt,
  ploid,
  snpsubset = NULL,
  method = "VanRaden",
  phat,
  ep = 0,
  ...
)
}
\arguments{
\item{method}{Character string specifying whether the VanRaden (\code{'VanRaden'}) based estimator or 
the Weir-Goudet (\code{'WG'}) estimator is used to construct the GRM.}

\item{filter}{Named list of thresholds for various filtering criteria.
See below for details.}
}
\description{
Method for constructing a genomic relationship matrix (GRM) for a diploid or autopolyploid population.
}
\seealso{
\code{\link{GRM}}
}
\author{
Timothy P. Bilton
}
