% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGRM.R
\name{makeGRM}
\alias{makeGRM}
\title{Make an genomic relationship matix (GRM) object}
\usage{
makeGRM(RAobj, ploid = 2, method = "VanRaden", indsubset = NULL,
  nThreads = 1, ep = 0, filter = list(MAF = NULL, MISS = NULL, PVALUE
  = NULL), est = list(MAF = TRUE, HWE = TRUE))
}
\value{
An R6 object of class GRM.
}
\description{
Make an genomic relationship matix (GRM) object
}
