% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRM.R
\name{GRM}
\alias{GRM}
\title{GRM object}
\description{
Class for storing RA data and associated functions for analyzing unstructured populations (e.g.,
populations with no known structure).
}
\details{
A genomic relationship matrix (GRM) object is created from the \code{\link{makeGRM}} function and contains RA data,
various statistics of the dataset that have been computed, and functions (or methods)
for analyzing the data. Information in a GRM object are specific to constructing a genomic relationship matrix
}
\seealso{
\code{\link{makeGRM}}
}
\author{
Timothy P. Bilton
}
\section{Super class}{
\code{\link[GUSbase:RA]{GUSbase::RA}} -> \code{GRM}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GRM-new}{\code{GRM$new()}}
\item \href{#method-GRM-print}{\code{GRM$print()}}
\item \href{#method-GRM-computeGRM}{\code{GRM$computeGRM()}}
\item \href{#method-GRM-HWEtest}{\code{GRM$HWEtest()}}
\item \href{#method-GRM-removeGRM}{\code{GRM$removeGRM()}}
\item \href{#method-GRM-PCA}{\code{GRM$PCA()}}
\item \href{#method-GRM-addSampleInfo}{\code{GRM$addSampleInfo()}}
\item \href{#method-GRM-removeSampleInfo}{\code{GRM$removeSampleInfo()}}
\item \href{#method-GRM-extractGRM}{\code{GRM$extractGRM()}}
\item \href{#method-GRM-writeGRM}{\code{GRM$writeGRM()}}
\item \href{#method-GRM-clone}{\code{GRM$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="GUSbase" data-topic="RA" data-id="RDDPlot"><a href='../../GUSbase/html/RA.html#method-RA-RDDPlot'><code>GUSbase::RA$RDDPlot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="GUSbase" data-topic="RA" data-id="cometPlot"><a href='../../GUSbase/html/RA.html#method-RA-cometPlot'><code>GUSbase::RA$cometPlot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="GUSbase" data-topic="RA" data-id="extractVar"><a href='../../GUSbase/html/RA.html#method-RA-extractVar'><code>GUSbase::RA$extractVar()</code></a></span></li>
<li><span class="pkg-link" data-pkg="GUSbase" data-topic="RA" data-id="mergeSamples"><a href='../../GUSbase/html/RA.html#method-RA-mergeSamples'><code>GUSbase::RA$mergeSamples()</code></a></span></li>
<li><span class="pkg-link" data-pkg="GUSbase" data-topic="RA" data-id="rocketPlot"><a href='../../GUSbase/html/RA.html#method-RA-rocketPlot'><code>GUSbase::RA$rocketPlot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="GUSbase" data-topic="RA" data-id="writeRA"><a href='../../GUSbase/html/RA.html#method-RA-writeRA'><code>GUSbase::RA$writeRA()</code></a></span></li>
<li><span class="pkg-link" data-pkg="GUSbase" data-topic="RA" data-id="writeVCF"><a href='../../GUSbase/html/RA.html#method-RA-writeVCF'><code>GUSbase::RA$writeVCF()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GRM-new"></a>}}
\if{latex}{\out{\hypertarget{method-GRM-new}{}}}
\subsection{Method \code{new()}}{
Method for initializing GRM object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GRM$new(RAobj, ploid, indsubset, saminfo)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{RAobj}}{Existing RA object to use in initializing the GRM object.}

\item{\code{ploid}}{Vector of integer values indicating the ploid levels of each individual.}

\item{\code{indsubset}}{Vector of indices specifying which individuals from the RA object to retain.}

\item{\code{saminfo}}{Data frame of sample information.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GRM-print"></a>}}
\if{latex}{\out{\hypertarget{method-GRM-print}{}}}
\subsection{Method \code{print()}}{
Print output from a GRM object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GRM$print()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Prints the summary information of the dataset and if any GRMs have been constructed, it 
also prints out summary information of the GRM runs.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GRM-computeGRM"></a>}}
\if{latex}{\out{\hypertarget{method-GRM-computeGRM}{}}}
\subsection{Method \code{computeGRM()}}{
Construct a genomic relationship matrix (GRM)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GRM$computeGRM(
  name,
  method = "VanRaden",
  ep = 0,
  snpsubset = NULL,
  filter = list(MAF = NULL, MISS = NULL, PVALUE = NULL),
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Specific name given to the run constructing the GRM.}

\item{\code{method}}{Character value indicating which method to used to construct the GRM.}

\item{\code{ep}}{Sequencing error rate. Can be a single value, a vector equal to the number of SNPs
or a matrix the same dimension as the data}

\item{\code{snpsubset}}{Vector of indices indicated which SNPs to retain to construct the GRM.}

\item{\code{filter}}{Name list specifying the filtering criteria to be applied.}

\item{\code{...}}{Not used currently.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The filtering criteria currently implemented are:
\itemize{
\item{Minor allele frequency (\code{MAF}): }{SNPs are discarded if their MAF is less than the threshold (default is \code{NULL})}
\item{Proportion of missing data (\code{MISS}): }{SNPs are discarded if the proportion of individuals with no reads (e.g. missing genotype)
 is greater than the threshold value (default is \code{NULL}).}
\item{P-value from a HWE test(\code{PVALUE}):}{ SNPs are discarded if the p-value from the Hardy-Weinberg equilibrium test is less than
the threshold. (default is \code{NULL})}
}
If a filtering criteria is set to \code{NULL}, then no filtering in regard to
that threshold is applied.

Not that for the \code{PVALUE} filter, the HWE test must first be run using the \code{$HWEtest} method.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GRM-HWEtest"></a>}}
\if{latex}{\out{\hypertarget{method-GRM-HWEtest}{}}}
\subsection{Method \code{HWEtest()}}{
Perform a Hardy–Weinberg Equilibrium (HWE) test for autopolyploids.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GRM$HWEtest(
  snpsubset = NULL,
  indsubset = NULL,
  nThreads = 1,
  para = NULL,
  EMpara = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{snpsubset}}{Vector of SNP indices indicating which SNPs to do the HWE test on.
Excluded SNPs are given a p-value of 0.}

\item{\code{indsubset}}{Vector of indices of the samples indicating which samples to use in the HWE test.}

\item{\code{nThreads}}{Integer value specifying the number of threads to use in the parallelization.}

\item{\code{para}}{Starting values passed to the \code{\link[GUSbase]{p_est_em}} and \code{\link[GUSbase]{g_est_em}} functions.}

\item{\code{EMpara}}{Convergence criteria passed to the \code{\link[GUSbase]{p_est_em}} and \code{\link[GUSbase]{g_est_em}} functions.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The function calls the \code{\link[GUSbase]{p_est_em}} and \code{\link[GUSbase]{g_est_em}} functions in the 
\code{\link[GUSbase]{GUSbase}} package to perform the HWE test.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GRM-removeGRM"></a>}}
\if{latex}{\out{\hypertarget{method-GRM-removeGRM}{}}}
\subsection{Method \code{removeGRM()}}{
Delete a GRM run within the GRM object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GRM$removeGRM(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the GRM run to delete.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GRM-PCA"></a>}}
\if{latex}{\out{\hypertarget{method-GRM-PCA}{}}}
\subsection{Method \code{PCA()}}{
Produce a PCA plot of a constructed GRM
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GRM$PCA(
  name,
  npc = 3,
  colour = NULL,
  shape = NULL,
  group.hover = NULL,
  interactive = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Character value giving the name of the GRM run to extract the constructed GRM from.}

\item{\code{npc}}{Integer number specifying the number of PCs to plot}

\item{\code{colour}}{Character string specifying which column of the sample information to use to colour the points}

\item{\code{shape}}{Character string specifying which column of the sample information to use to group the points based on point shape}

\item{\code{group.hover}}{Character vector specifying which column(s) of the sample information to include in the hover information for the
interactive plots. Only used if \code{interactive=TRUE}.}

\item{\code{interactive}}{Logical value: If \code{TRUE}, then an interactive plot using the \code{plotly} package is produced.
Otherwise, a standard \code{ggplot} is constructed.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The sample information used by this function is the information added by the `$addSampleInfo` method.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GRM-addSampleInfo"></a>}}
\if{latex}{\out{\hypertarget{method-GRM-addSampleInfo}{}}}
\subsection{Method \code{addSampleInfo()}}{
Add sample information to the GRM object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GRM$addSampleInfo(samfile)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{samfile}}{Path to the file that contains the sample information to be loaded to GRM object.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The sample information must contain a column 'ID' that has the ID of the samples and must match
the IDs of the samples already in the GRM object. Any extra columns will be added to the existing 
sample information data.

Note that a column named 'Ploidy' can not be present in the file. If there are any columns already present
then the function throw and error.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GRM-removeSampleInfo"></a>}}
\if{latex}{\out{\hypertarget{method-GRM-removeSampleInfo}{}}}
\subsection{Method \code{removeSampleInfo()}}{
Remove sample information from a GRM object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GRM$removeSampleInfo(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Character vector of column names to remove from the sample information data.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Note that columns 'ID' and 'Ploidy' cannot be deleted.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GRM-extractGRM"></a>}}
\if{latex}{\out{\hypertarget{method-GRM-extractGRM}{}}}
\subsection{Method \code{extractGRM()}}{
Extract a GRM from a run.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GRM$extractGRM(name, IDvar = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Character value specifying which run to extract the GRM from.}

\item{\code{IDvar}}{Character value specifying which column of the sample information to use for 
IDs in the GRM matrix. If \code{IDvar=NULL}, then the 'ID' column is used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GRM-writeGRM"></a>}}
\if{latex}{\out{\hypertarget{method-GRM-writeGRM}{}}}
\subsection{Method \code{writeGRM()}}{
Write GRM to file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GRM$writeGRM(name, filename, IDvar = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Character value specifying which run to write the GRM from.}

\item{\code{filename}}{Character value giving the name of the output file.}

\item{\code{IDvar}}{Character value specifying which column of the sample information to use for 
IDs in the GRM matrix. If \code{IDvar=NULL}, then the 'ID' column is used.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The file format is currently a symmetric matrix with the IDs on the rows and columns.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GRM-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GRM-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GRM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
