% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRM.R
\name{GRM}
\alias{GRM}
\title{GRM object}
\description{
Class for storing RA data and associated functions for analyzing unstructured populations (e.g.,
populations with no known structure).
}
\details{
A genomic relationship matrix (GRM) object is created from the \code{\link{makeGRM}} function and contains RA data,
various statistics of the dataset that have been computed, and functions (or methods)
for analyzing the data. Information in an GRM are specific to constructing a GRM
}
\seealso{
\code{\link{makeGRM}}
}
\author{
Timothy P. Bilton
}
\section{Super class}{
\code{\link[GUSbase:RA]{GUSbase::RA}} -> \code{GRM}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GRM-new}{\code{GRM$new()}}
\item \href{#method-GRM-print}{\code{GRM$print()}}
\item \href{#method-GRM-computeGRM}{\code{GRM$computeGRM()}}
\item \href{#method-GRM-HWEtest}{\code{GRM$HWEtest()}}
\item \href{#method-GRM-removeGRM}{\code{GRM$removeGRM()}}
\item \href{#method-GRM-plotGRM}{\code{GRM$plotGRM()}}
\item \href{#method-GRM-CompareGRM}{\code{GRM$CompareGRM()}}
\item \href{#method-GRM-PCA}{\code{GRM$PCA()}}
\item \href{#method-GRM-addSampleInfo}{\code{GRM$addSampleInfo()}}
\item \href{#method-GRM-removeSampleInfo}{\code{GRM$removeSampleInfo()}}
\item \href{#method-GRM-extractGRM}{\code{GRM$extractGRM()}}
\item \href{#method-GRM-writeGRM}{\code{GRM$writeGRM()}}
\item \href{#method-GRM-clone}{\code{GRM$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="GUSbase" data-topic="RA" data-id="RDDPlot"><a href='../../GUSbase/html/RA.html#method-RA-RDDPlot'><code>GUSbase::RA$RDDPlot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="GUSbase" data-topic="RA" data-id="cometPlot"><a href='../../GUSbase/html/RA.html#method-RA-cometPlot'><code>GUSbase::RA$cometPlot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="GUSbase" data-topic="RA" data-id="extractVar"><a href='../../GUSbase/html/RA.html#method-RA-extractVar'><code>GUSbase::RA$extractVar()</code></a></span></li>
<li><span class="pkg-link" data-pkg="GUSbase" data-topic="RA" data-id="mergeSamples"><a href='../../GUSbase/html/RA.html#method-RA-mergeSamples'><code>GUSbase::RA$mergeSamples()</code></a></span></li>
<li><span class="pkg-link" data-pkg="GUSbase" data-topic="RA" data-id="rocketPlot"><a href='../../GUSbase/html/RA.html#method-RA-rocketPlot'><code>GUSbase::RA$rocketPlot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="GUSbase" data-topic="RA" data-id="writeRA"><a href='../../GUSbase/html/RA.html#method-RA-writeRA'><code>GUSbase::RA$writeRA()</code></a></span></li>
<li><span class="pkg-link" data-pkg="GUSbase" data-topic="RA" data-id="writeVCF"><a href='../../GUSbase/html/RA.html#method-RA-writeVCF'><code>GUSbase::RA$writeVCF()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GRM-new"></a>}}
\if{latex}{\out{\hypertarget{method-GRM-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GRM$new(GRMobj, ploid, indsubset, saminfo)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GRM-print"></a>}}
\if{latex}{\out{\hypertarget{method-GRM-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GRM$print(what = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GRM-computeGRM"></a>}}
\if{latex}{\out{\hypertarget{method-GRM-computeGRM}{}}}
\subsection{Method \code{computeGRM()}}{
Construct a Genomic relationship matrix (GRM)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GRM$computeGRM(
  name,
  method = "VanRaden",
  ep = 0,
  snpsubset = NULL,
  filter = list(MAF = NULL, MISS = NULL, PVALUE = NULL),
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Specific name given to the run constructing the GRM.}

\item{\code{method}}{Character value indicating which method to used to contruct the GRM.}

\item{\code{ep}}{Sequencing error rate. Can be a single value, a vector equal to the number of SNPs
or a matrix the same dimension as the data}

\item{\code{snpsubset}}{Vector of indices indicated which SNPs to retain to contruct the GRM.}

\item{\code{filter}}{Name list specifying the filtering criteria to be applied.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GRM-HWEtest"></a>}}
\if{latex}{\out{\hypertarget{method-GRM-HWEtest}{}}}
\subsection{Method \code{HWEtest()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GRM$HWEtest(
  snpsubset = NULL,
  indsubset = NULL,
  nThreads = 1,
  para = NULL,
  EMpara = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GRM-removeGRM"></a>}}
\if{latex}{\out{\hypertarget{method-GRM-removeGRM}{}}}
\subsection{Method \code{removeGRM()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GRM$removeGRM(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GRM-plotGRM"></a>}}
\if{latex}{\out{\hypertarget{method-GRM-plotGRM}{}}}
\subsection{Method \code{plotGRM()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GRM$plotGRM(
  type = c("histrogram", "scatter"),
  values = c("both", "self-relatedness", "relatednes"),
  xaxis = NULL,
  interactive = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GRM-CompareGRM"></a>}}
\if{latex}{\out{\hypertarget{method-GRM-CompareGRM}{}}}
\subsection{Method \code{CompareGRM()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GRM$CompareGRM()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GRM-PCA"></a>}}
\if{latex}{\out{\hypertarget{method-GRM-PCA}{}}}
\subsection{Method \code{PCA()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GRM$PCA(
  name,
  npc = 3,
  colour = NULL,
  shape = NULL,
  group.hover = NULL,
  interactive = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GRM-addSampleInfo"></a>}}
\if{latex}{\out{\hypertarget{method-GRM-addSampleInfo}{}}}
\subsection{Method \code{addSampleInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GRM$addSampleInfo(samfile)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GRM-removeSampleInfo"></a>}}
\if{latex}{\out{\hypertarget{method-GRM-removeSampleInfo}{}}}
\subsection{Method \code{removeSampleInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GRM$removeSampleInfo(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GRM-extractGRM"></a>}}
\if{latex}{\out{\hypertarget{method-GRM-extractGRM}{}}}
\subsection{Method \code{extractGRM()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GRM$extractGRM(name, IDvar = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GRM-writeGRM"></a>}}
\if{latex}{\out{\hypertarget{method-GRM-writeGRM}{}}}
\subsection{Method \code{writeGRM()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GRM$writeGRM(name, filename, IDvar = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GRM-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GRM-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GRM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
