% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUSrelate.R
\docType{data}
\name{GRM}
\alias{GRM}
\title{GRM object}
\usage{
USobj <- makeUS()
}
\description{
Class for storing RA data and associated functions for analyzing unstructured populations (e.g.,
populations with no known structure).
}
\details{
An US object is created from the \code{\link{makeUS}} function and contains RA data,
various statistics of the dataset that have been computed, and functions (or methods)
for analyzing the data. Information in an US are specific to unrelated populations (or
populations with no known relationships).
}
\seealso{
\code{\link{makeUR}}
}
\author{
Timothy P. Bilton
}
\keyword{datasets}
