% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUSrelate-package.R
\docType{package}
\name{GUSbase-package}
\alias{GUSbase-package}
\alias{_PACKAGE}
\alias{GUSrelate}
\title{Genotyping Uncertainty with Sequencing data and RELATEdness (GUSrelate)}
\description{
GUSrelate is a package for computing relatedness between all pairs of individuals in a population
via reference and alternate count information from sequence data.
}
\details{
\tabular{ll}{ Package: \tab GUSrelate\cr Type: \tab Package\cr Version: \tab
0.2.2\cr Date: \tab 2024-01-24\cr License: \tab GPL-3\cr }
}
\references{
\insertAllCited{}
}
\author{
Timothy P. Bilton, Maintainer: Timothy P. Bilton
<timothy.bilton@agresearch.ac.nz>
}
\keyword{package}
